+++ realm/realm-annotations-processor/src/test/resources/some/test/EmbeddedClass.java
@@ -1,0 +1,11 @@
+package some.test;
+
+import io.realm.RealmObject;
+import io.realm.RealmResults;
+import io.realm.annotations.RealmClass;
+
+@RealmClass(embedded = true)
+public class EmbeddedClass extends RealmObject {
+    public String name;
+    public int age;
+}